requirements.txt
python-dotenv
minio
neo4j
langchain
langchain-community
langchain-openai
python-docx
pypdf
tiktoken
